ProductList tamamlandı

+++ Entity Layer/Product.cs
@@ -11,7 +11,7 @@ namespace Entity_Layer
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
-        public float UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int StockQuantity { get; set; }
         public int CategoryId { get; set; } 
         public Category Category { get; set; }
